requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
pypdf
duckduckgo-search
python-dotenv
